start consolidating json utils
