--- server/servers/api/src/main/scala/com/prisma/util/json/JsonFormats.scala
@@ -1,67 +1,0 @@
-package com.prisma.util.json
-
-import play.api.libs.json._
-
-object JsonFormats {
-
-//  implicit object CaseClassFormat extends JsonFormat[Product] {
-//    def write(x: Product): JsValue = {
-//      val values = x.productIterator.toList
-//      val fields = x.getClass.getDeclaredFields
-//
-//      def getIdValue(p: Product): Option[Any] = {
-//        val values = p.productIterator.toList
-//        val fields = p.getClass.getDeclaredFields
-//
-//        fields.zipWithIndex.find(_._1.getName == "id").map(z => values(z._2))
-//      }
-//
-//      val map: Map[String, Any] = values.zipWithIndex.map {
-//        case (v, i) =>
-//          val key = fields(i).getName
-//          val value = v match {
-//            case v: Product if !v.isInstanceOf[Option[_]] => getIdValue(v).getOrElse("...")
-//            case Some(v: Product)                         => getIdValue(v).getOrElse("...")
-//            case v                                        => v
-//          }
-//
-//          key -> value
-//      }.toMap
-//
-//      AnyJsonFormat.write(map)
-//    }
-//
-//    def read(value: JsValue) = throw new UnsupportedOperationException()
-//  }
-
-  implicit object AnyJsonFormat extends Writes[Any] {
-    def writes(x: Any): JsValue = x match {
-      case m: Map[_, _] => JsObject(m.asInstanceOf[Map[String, Any]].mapValues(writes))
-      case l: List[Any] => JsArray(l.map(writes).toVector)
-      case n: Int       => JsNumber(n)
-      case n: Long      => JsNumber(n)
-      case n: Double    => JsNumber(n)
-      case s: String    => JsString(s)
-      case true         => JsTrue
-      case false        => JsFalse
-      case v: JsValue   => v
-      case null         => JsNull
-      case r            => JsString(r.toString)
-    }
-  }
-
-  implicit object MapJsonWriter extends Writes[Map[String, Any]] {
-    override def writes(obj: Map[String, Any]): JsValue = AnyJsonFormat.writes(obj)
-  }
-
-//  class SeqAnyJsonWriter[T <: Any] extends Writes[Seq[Map[String, T]]] {
-//    override def write(objs: Seq[Map[String, T]]): JsValue =
-//      new JsArray(
-//        objs
-//          .map(obj => {
-//            AnyJsonFormat.write(obj)
-//          })
-//          .toVector)
-//  }
-
-}
